+++ BBMOD_CLI/include/BBMOD/Node.hpp
@@ -26,7 +26,7 @@ struct SNode
 
 	dual_quat_t Transform;
 
-	std::vector<size_t> Meshes;
+	std::vector<uint32_t> Meshes;
 
 	std::vector<SNode*> Children;
 };
