+++ BBMOD_CLI/src/BBMOD/Importer.cpp
@@ -50,9 +50,9 @@ static std::string GetAnimationFilename(SAnimation* animation, int index, const 
 	return GetFilename(out, animationName.c_str(), ".bbanim");
 }
 
-static void LogNode(std::ofstream& log, SNode* node, size_t indent)
+static void LogNode(std::ofstream& log, SNode* node, uint32_t indent)
 {
-	for (size_t i = 0; i < indent * 4; ++i)
+	for (uint32_t i = 0; i < indent * 4; ++i)
 	{
 		log << " ";
 	}
@@ -175,7 +175,7 @@ int ConvertToBBMOD(const char* fin, const char* fout, const SConfig& config)
 	if (model->BoneCount > 64)
 	{
 		PRINT_WARNING(
-			"This model has %zd bones, but the default upper limit defined in shader BBMOD_ShDefaultAnimated is 64!"
+			"This model has %d bones, but the default upper limit defined in shader BBMOD_ShDefaultAnimated is 64!"
 			" You will need to increase this limit in order to render this model, though be aware that the maximum"
 			" number of vertex shader uniforms is determined by the target platform! Setting it higher than 64 can"
 			" make your game incompatible with some devices!"
@@ -191,7 +191,7 @@ int ConvertToBBMOD(const char* fin, const char* fout, const SConfig& config)
 
 	log << "Materials:" << std::endl;
 	log << "==========" << std::endl;
-	for (size_t i = 0; i < model->MaterialNames.size(); ++i)
+	for (uint32_t i = 0; i < model->MaterialNames.size(); ++i)
 	{
 		log << i << ": " << model->MaterialNames[i] << std::endl;
 	}
